Heavy refactoring to make the init file more modular
Things are now moved around better. Still not totally done yet, but
this structure and flow is much more manageable.

Elpa still doesn't want to load files into the elpa-25.3 files yet,
so there might be an extra use-package config or something to set, but
this is a really good start