Make changes to how packages are loaded
Custom config file was blowing up if it loaded too early due to a
config option that references `projectile` if projectile doesn't exist
yet.

Also detects if packages weren't installed yet by if use-package
doesn't exist, and then refreshes packages only then. This allows the
config to be independent of if the package repositories can be
reached, allowing offline usage