Add lsp-mode and configure to use with go-mode
the underlying lsp server is implemented by `gopls`. The installation
instructions are listed in the file. A next good move would be to
either make a script for my emacs config to prepare a system to use
with the emacs config, or build out a Docker container that can be
used by lsp to download and run. Then the only dependency for lsp-mode
would be for the native system to have docker available